// src/components/keyboard/Keyboard.types.ts
export type Status = "misplaced" | "incorrect" | "correct" | "gram" | "";

export type KeysStatus = Record<string, Status>;

export type Key =
  | "Q"
  | "W"
  | "E"
  | "R"
  | "T"
  | "Y"
  | "U"
  | "I"
  | "O"
  | "P"
  | "A"
  | "S"
  | "D"
  | "F"
  | "G"
  | "H"
  | "J"
  | "K"
  | "L"
  | "Z"
  | "X"
  | "C"
  | "V"
  | "B"
  | "N"
  | "M"
  | "`"
  | "Backspace"
  | "Enter"
  | "Blank"
  | "TAB"
  | "ESCAPE"
  | "Gram"
  | " "
  | "Space"
  | "spacer"
  | "end";